pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adminlineage"
version = "0.1.0"
description = "Build administrative evolution keys across time with anchor-constrained Gemini adjudication"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "AdminLineage Contributors" }
]
dependencies = [
  "pandas>=2.0",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "google-genai>=0.7",
]

[project.optional-dependencies]
io = ["pyarrow>=15.0"]
dev = [
  "pytest>=8.0",
  "ruff>=0.5",
]

[project.scripts]
adminlineage = "adminlineage.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
